block buffer fusing

+++ src/include/deconvolution_inst.h
@@ -73,7 +73,7 @@ public:
         if (static_cast<int32_t>(idx) > 0)
             throw std::range_error("Only one input for fused sum is supported");
 
-        int d_idx = 1 + this->get_split() + idx;
+        size_t d_idx = 1 + this->get_split() + idx;
         d_idx += bias_term() ? this->get_split() : 0;
         return get_dependency(d_idx);
     }
